refactor: remove ExternalLink component and update imports in explore.tsx for cleaner structure
